+++ src/main/java/org/mattlang/jc/ConfigValues.java
@@ -90,6 +90,10 @@ public class ConfigValues {
     public final UCICheckOption expandPv = internal.createCheckOpt("expandPv",
             "should the found PV expand by cache entries? Otherwise they could be shorter than the depth caused by pruning.",
             true);
+    public final UCICheckOption mateDistancePruning = internal.createCheckOpt("mateDistancePruning",
+            "should mate distance pruning be activated?",
+            true);
+
 
     public final UCICheckOption aspiration = pruning.createCheckOpt("aspiration",
             "should aspiration windows be used during iterative deepening",